Complete redesign of themepage usage.
The themepage itself contains the bare minimum of mnw-specific stuff. It should
not need changes on mnw updates anymore, but only be a wrapper around mnw in
order to style the page according to the theme.

mnw_parse_request and the remainings of the old themepage are now merged in
the class mnw_Themepage.

Translation of the themepage is now realized with gettext.

These changes make upgrades easier, reduces load on pages other than the theme
page (since mnw.php is smaller now) and simplifies translations.